Update based on unit tests
