+++ lib/js/src/manager/screen/choiceset/ChoiceCell.js
@@ -246,6 +246,9 @@ class ChoiceCell {
         if (this.getTertiaryText() !== other.getTertiaryText()) {
             return false;
         }
+        if (this._getUniqueText() !== other._getUniqueText()) {
+            return false;
+        }
         if (this.getArtwork() === null && other.getArtwork() !== null) {
             return false;
         }
