+++ lib/js/src/manager/screen/choiceset/_ChoiceSetManagerBase.js
@@ -151,25 +151,6 @@ class _ChoiceSetManagerBase extends _SubManagerBase {
     }
 
     /**
-     * Checks if 2 or more cells have the same text/title. In case this condition is true, this function will handle the presented issue by adding "(count)".
-     * E.g. Choices param contains 2 cells with text/title "Address" will be handled by updating the uniqueText/uniqueTitle of the second cell to "Address (2)".
-     * @param {ChoiceCell[]} choices - A list of ChoiceCell objects to be uploaded
-     */
-    _addUniqueNamesToCells (choices) {
-        const dictCounter = {}; // create a string to number hash for counting similar primary texts
-
-        choices.forEach(choice => {
-            const choiceName = choice.getText();
-            if (dictCounter[choiceName] === undefined) {
-                dictCounter[choiceName] = 1; // unique text
-            } else { // found a duplicate
-                dictCounter[choiceName] += 1;
-                choice._setUniqueText(`${choiceName} (${dictCounter[choiceName]})`);
-            }
-        });
-    }
-
-    /**
      * Deletes choices that were sent previously
      * @param {ChoiceCell[]} choices - A list of ChoiceCell objects
      * @returns {Promise} - A promise that resolves to a Boolean of whether the operation is a success
@@ -470,32 +451,6 @@ class _ChoiceSetManagerBase extends _SubManagerBase {
         }
 
         return modifiedKeyboardConfiguration;
-    }
-
-    /**
-     * Finds non-unique choice cells and updates their unique text accordingly
-     * @param {ChoiceCell[]} strippedCells - Choice cells with their unsupported properties removed
-     * @param {ChoiceCell[]} unstrippedCells - The original choice cells
-     */
-    _addUniqueNamesBasedOnStrippedCells (strippedCells, unstrippedCells) {
-        if (!Array.isArray(strippedCells) || !Array.isArray(unstrippedCells) || strippedCells.length !== unstrippedCells.length) {
-            return;
-        }
-        // array of unique choice cells
-        const cells = [];
-        // array of the count of how many times each cell has been found
-        const cellsCounter = [];
-        strippedCells.forEach((strippedCell, index) => {
-            // find if a previous cell was a duplicate and update unique text of the current cell if so
-            const duplicateIndex = cells.map(cell => cell.equals(strippedCell)).indexOf(true);
-            if (duplicateIndex >= 0) {
-                cellsCounter[duplicateIndex]++;
-                unstrippedCells[index]._setUniqueText(`${unstrippedCells[index].getText()} (${cellsCounter[duplicateIndex]})`);
-            } else {
-                cells.push(strippedCell);
-                cellsCounter.push(1);
-            }
-        });
     }
