+++ lib/js/src/manager/screen/choiceset/_PreloadPresentChoicesOperation.js
@@ -68,7 +68,7 @@ class _PreloadPresentChoicesOperation extends _Task {
      */
     constructor (lifecycleManager = null, fileManager = null, displayName = null,
         defaultMainWindowCapability = null, isVrOptional, cellsToPreload = [],
-        loadedCells = [], choiceSet, presentationMode = null,
+        loadedCells = [], choiceSet = null, presentationMode = null,
         originalKeyboardProperties, keyboardListener = null,
         choiceSetSelectionListener, cancelId, completionListener = null) {
         super('_PreloadPresentChoicesOperation');
@@ -182,6 +182,51 @@ class _PreloadPresentChoicesOperation extends _Task {
     }
 
     /**
+     * Checks if 2 or more cells have the same text/title. In case this condition is true, this function will handle the presented issue by adding "(count)".
+     * E.g. Choices param contains 2 cells with text/title "Address" will be handled by updating the uniqueText/uniqueTitle of the second cell to "Address (2)".
+     * @param {ChoiceCell[]} choices - A list of ChoiceCell objects to be uploaded
+     */
+    _addUniqueNamesToCells (choices) {
+        const dictCounter = {}; // create a string to number hash for counting similar primary texts
+
+        choices.forEach(choice => {
+            const choiceName = choice.getText();
+            if (dictCounter[choiceName] === undefined) {
+                dictCounter[choiceName] = 1; // unique text
+            } else { // found a duplicate
+                dictCounter[choiceName] += 1;
+                choice._setUniqueText(`${choiceName} (${dictCounter[choiceName]})`);
+            }
+        });
+    }
+
+    /**
+     * Finds non-unique choice cells and updates their unique text accordingly
+     * @param {ChoiceCell[]} strippedCells - Choice cells with their unsupported properties removed
+     * @param {ChoiceCell[]} unstrippedCells - The original choice cells
+     */
+    _addUniqueNamesBasedOnStrippedCells (strippedCells, unstrippedCells) {
+        if (!Array.isArray(strippedCells) || !Array.isArray(unstrippedCells) || strippedCells.length !== unstrippedCells.length) {
+            return;
+        }
+        // array of unique choice cells
+        const cells = [];
+        // array of the count of how many times each cell has been found
+        const cellsCounter = [];
+        strippedCells.forEach((strippedCell, index) => {
+            // find if a previous cell was a duplicate and update unique text of the current cell if so
+            const duplicateIndex = cells.map(cell => cell.equals(strippedCell)).indexOf(true);
+            if (duplicateIndex >= 0) {
+                cellsCounter[duplicateIndex]++;
+                unstrippedCells[index]._setUniqueText(`${unstrippedCells[index].getText()} (${cellsCounter[duplicateIndex]})`);
+            } else {
+                cells.push(strippedCell);
+                cellsCounter.push(1);
+            }
+        });
+    }
+
+    /**
      * Modifies the choices names depending on SDL version
      * @param {ChoiceCell[]} choicesToUpload - The first list of choices
      * @returns {ChoiceCell[]} - A deep copy of the name modified choices
@@ -201,9 +246,9 @@ class _PreloadPresentChoicesOperation extends _Task {
 
         return choicesClone.filter((clonedChoice) => {
             // returns false if the cloned choice appears in the list of preloaded choices
-            return !this._preloadedChoices.map((preloadedChoice) => {
+            return !this._loadedCells.map((loadedCell) => {
                 // the unique text is important for this comparison but it isn't checked by .equals()
-                return clonedChoice.equals(preloadedChoice) && (clonedChoice._getUniqueText() === preloadedChoice._getUniqueText());
+                return clonedChoice.equals(loadedCell) && (clonedChoice._getUniqueText() === loadedCell._getUniqueText());
             }).includes(true);
         });
     }
@@ -702,7 +747,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for NOT_STARTED.
      * @returns {String} - The enum value.
      */
-    static NOT_STARTED () {
+    static get NOT_STARTED () {
         return _PreloadPresentChoicesOperationState._MAP.NOT_STARTED;
     }
 
@@ -710,7 +755,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for UPLOADING_IMAGES.
      * @returns {String} - The enum value.
      */
-    static UPLOADING_IMAGES () {
+    static get UPLOADING_IMAGES () {
         return _PreloadPresentChoicesOperationState._MAP.UPLOADING_IMAGES;
     }
 
@@ -718,7 +763,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for UPLOADING_CHOICES.
      * @returns {String} - The enum value.
      */
-    static UPLOADING_CHOICES () {
+    static get UPLOADING_CHOICES () {
         return _PreloadPresentChoicesOperationState._MAP.UPLOADING_CHOICES;
     }
 
@@ -726,7 +771,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for UPDATING_KEYBOARD_PROPERTIES.
      * @returns {String} - The enum value.
      */
-    static UPDATING_KEYBOARD_PROPERTIES () {
+    static get UPDATING_KEYBOARD_PROPERTIES () {
         return _PreloadPresentChoicesOperationState._MAP.UPDATING_KEYBOARD_PROPERTIES;
     }
 
@@ -734,7 +779,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for PRESENTING_CHOICES.
      * @returns {String} - The enum value.
      */
-    static PRESENTING_CHOICES () {
+    static get PRESENTING_CHOICES () {
         return _PreloadPresentChoicesOperationState._MAP.PRESENTING_CHOICES;
     }
 
@@ -742,7 +787,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for CANCELLING_PRESENT_CHOICES.
      * @returns {String} - The enum value.
      */
-    static CANCELLING_PRESENT_CHOICES () {
+    static get CANCELLING_PRESENT_CHOICES () {
         return _PreloadPresentChoicesOperationState._MAP.CANCELLING_PRESENT_CHOICES;
     }
 
@@ -750,7 +795,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for RESETTING_KEYBOARD_PROPERTIES.
      * @returns {String} - The enum value.
      */
-    static RESETTING_KEYBOARD_PROPERTIES () {
+    static get RESETTING_KEYBOARD_PROPERTIES () {
         return _PreloadPresentChoicesOperationState._MAP.RESETTING_KEYBOARD_PROPERTIES;
     }
 
@@ -758,7 +803,7 @@ class _PreloadPresentChoicesOperationState extends Enum {
      * Get the enum value for FINISHING.
      * @returns {String} - The enum value.
      */
-    static FINISHING () {
+    static get FINISHING () {
         return _PreloadPresentChoicesOperationState._MAP.FINISHING;
     }
 }
@@ -773,5 +818,7 @@ _PreloadPresentChoicesOperationState._MAP = Object.freeze({
     'RESETTING_KEYBOARD_PROPERTIES': 'RESETTING_KEYBOARD_PROPERTIES',
     'FINISHING': 'FINISHING',
 });
+
+_PreloadPresentChoicesOperation._PreloadPresentChoicesOperationState = _PreloadPresentChoicesOperationState;
 
 export { _PreloadPresentChoicesOperation };
