+++ tests/managers/screen/choiceset/ChoiceSetManagerTests.js
@@ -2,7 +2,6 @@ const SDL = require('../../../config.js').node;
 
 const Validator = require('../../../Validator');
 const sinon = require('sinon');
-const Test = require('../../../Test.js');
 
 module.exports = function (appClient) {
     describe('ChoiceSetManagerTests', function () {
@@ -288,115 +287,6 @@ module.exports = function (appClient) {
 
             // restore state
             csm._defaultMainWindowCapability = originCapability;
-        });
-
-        it('testAddUniqueNamesToCells', function () {
-            const cell1 = new SDL.manager.screen.choiceset.ChoiceCell('McDonalds')
-                .setSecondaryText('1 mile away');
-            const cell2 = new SDL.manager.screen.choiceset.ChoiceCell('McDonalds')
-                .setSecondaryText('2 miles away');
-            const cell3 = new SDL.manager.screen.choiceset.ChoiceCell('Starbucks')
-                .setSecondaryText('3 miles away');
-            const cell4 = new SDL.manager.screen.choiceset.ChoiceCell('McDonalds')
-                .setSecondaryText('4 miles away');
-            const cell5 = new SDL.manager.screen.choiceset.ChoiceCell('Starbucks')
-                .setSecondaryText('5 miles away');
-            const cell6 = new SDL.manager.screen.choiceset.ChoiceCell('Meijer')
-                .setSecondaryText('6 miles away');
-
-            csm._addUniqueNamesToCells([cell1, cell2, cell3, cell4, cell5, cell6]);
-
-            Validator.assertEquals(cell1._getUniqueText(), 'McDonalds');
-            Validator.assertEquals(cell2._getUniqueText(), 'McDonalds (2)');
-            Validator.assertEquals(cell3._getUniqueText(), 'Starbucks');
-            Validator.assertEquals(cell4._getUniqueText(), 'McDonalds (3)');
-            Validator.assertEquals(cell5._getUniqueText(), 'Starbucks (2)');
-            Validator.assertEquals(cell6._getUniqueText(), 'Meijer');
-        });
-
-        it('testUniquenessForAvailableFields', function () {
-            const windowCapability = new SDL.rpc.structs.WindowCapability();
-            const secondaryText = new SDL.rpc.structs.TextField()
-                .setNameParam(SDL.rpc.enums.TextFieldName.secondaryText);
-            const tertiaryText = new SDL.rpc.structs.TextField()
-                .setNameParam(SDL.rpc.enums.TextFieldName.tertiaryText);
-
-            const textFields = [
-                secondaryText,
-                tertiaryText,
-            ];
-            windowCapability.setTextFields(textFields);
-
-            const choiceImage = new SDL.rpc.structs.ImageField()
-                .setNameParam(SDL.rpc.enums.ImageFieldName.choiceImage);
-            const choiceSecondaryImage = new SDL.rpc.structs.ImageField()
-                .setNameParam(SDL.rpc.enums.ImageFieldName.choiceSecondaryImage);
-
-            const imageFieldList = [
-                choiceImage,
-                choiceSecondaryImage,
-            ];
-            windowCapability.setImageFields(imageFieldList);
-
-            csm._defaultMainWindowCapability = windowCapability;
-
-            const cell1 = new SDL.manager.screen.choiceset.ChoiceCell('Item 1')
-                .setSecondaryText('null')
-                .setTertiaryText('tertiaryText')
-                .setVoiceCommands(null)
-                .setArtwork(Test.GENERAL_ARTWORK)
-                .setSecondaryArtwork(Test.GENERAL_ARTWORK);
-            const cell2 = new SDL.manager.screen.choiceset.ChoiceCell('Item 1')
-                .setSecondaryText('null2')
-                .setTertiaryText('tertiaryText2')
-                .setVoiceCommands(null)
-                .setArtwork(null)
-                .setSecondaryArtwork(null);
-
-            const choiceCellList = [
-                cell1,
-                cell2,
-            ];
-
-            let removedProperties = csm._removeUnusedProperties(choiceCellList);
-            Validator.assertNotNullUndefined(removedProperties[0].getSecondaryText());
-
-            csm._defaultMainWindowCapability.setTextFields([]);
-            csm._defaultMainWindowCapability.setImageFields([]);
-
-            removedProperties = csm._removeUnusedProperties(choiceCellList);
-            csm._addUniqueNamesBasedOnStrippedCells(removedProperties, choiceCellList);
-            Validator.assertEquals(choiceCellList[1]._getUniqueText(), 'Item 1 (2)');
-        });
-
-        it('testChoicesToBeUploaded', function () {
-            const cell1 = new SDL.manager.screen.choiceset.ChoiceCell('Item 1')
-                .setSecondaryText('null')
-                .setTertiaryText('tertiaryText')
-                .setVoiceCommands(null)
-                .setArtwork(Test.GENERAL_ARTWORK)
-                .setSecondaryArtwork(Test.GENERAL_ARTWORK);
-            const cell2 = new SDL.manager.screen.choiceset.ChoiceCell('Item 2')
-                .setSecondaryText('null2')
-                .setTertiaryText('tertiaryText2')
-                .setVoiceCommands(null)
-                .setArtwork(null)
-                .setSecondaryArtwork(null);
-
-            const choiceCellList = [
-                cell1,
-                cell2,
-            ];
-
-            csm._preloadedChoices = choiceCellList;
-            Validator.assertEquals(csm._getChoicesToBeUploadedWithArray(choiceCellList), []);
-            const cell3 = new SDL.manager.screen.choiceset.ChoiceCell('Item 3')
-                .setSecondaryText('null3')
-                .setTertiaryText('tertiaryText3')
-                .setVoiceCommands(null)
-                .setArtwork(null)
-                .setSecondaryArtwork(null);
-            Validator.assertEquals(csm._getChoicesToBeUploadedWithArray([cell1, cell2, cell3]), [cell3]);
         });
     });
 };
